OMG GDM/Ubuntu/Unity
How to find out the current (active) XServer DISPLAY number?

Well, first, you may try to use
$ loginctl list-sessions

to get the list of sessions, and then use
$ loginctl show-session -p Display -p Active <session ID>

on each of these to get the X11 display number associated with
the currently active session.

...this doesn't work anymore with GDM, because in this case
the X server is started inside the session (with user
privileges) and /usr/bin/gdm-x-session does not communicate
this information back to logind (due to lack of an API for that).

I will try to find better solution of couse, but right now...
let's "enjoy" this bugfix. Sorry for that.